requirements.txt
# Основные зависимости
python-telegram-bot==20.7
openai==1.3.7
replicate==0.22.0
python-dotenv==1.0.0

# HTTP сервер для callback
Flask==3.0.0

# HTTP запросы
requests==2.31.0

# База данных
sqlite3

# Логирование
logging

# Время и дата
datetime 